Create Struct_AIPositions.uasset
Okay, so I have no idea if this commit/pull/push thing works, but I'm going to upload this struct that should work just like the wave struct, so it can be modified during run time if needed.
If I'm uploading this the wrong way, just tell me on teams.